+++ backend/canisters/user/impl/src/model/streak.rs
@@ -20,12 +20,21 @@ pub struct Streak {
 }
 
 impl Streak {
-    pub fn start_day(&self) -> u16 {
-        self.start_day
-    }
-
-    pub fn set_start_day(&mut self, start_day: u16) {
-        self.start_day = start_day;
+    // The timestamps should be the end of the day for which the claim applied to, as opposed to the
+    // start of the day after
+    pub fn fix_streak_claim_timestamps(&mut self) -> Vec<TimestampMillis> {
+        self.claims
+            .iter_mut()
+            .filter_map(|claim| {
+                let day = Self::timestamp_to_day(claim.timestamp).unwrap();
+                if claim.timestamp != Self::final_timestamp_of_day(day) {
+                    claim.timestamp = Self::final_timestamp_of_day(day - 1);
+                    Some(claim.timestamp)
+                } else {
+                    None
+                }
+            })
+            .collect()
     }
 
     pub fn days(&self, now: TimestampMillis) -> u16 {
@@ -77,7 +86,8 @@ impl Streak {
                 self.insurance_last_updated = now;
 
                 let claim = UserCanisterStreakInsuranceClaim {
-                    timestamp: now,
+                    // The timestamp of the end of the day for which the claim applied
+                    timestamp: Self::final_timestamp_of_day(today - 1),
                     streak_length: self.end_day - self.start_day,
                     new_days_claimed: self.days_missed,
                 };
@@ -176,6 +186,10 @@ impl Streak {
 
     pub fn day_to_timestamp(day: u16) -> TimestampMillis {
         DAY_ZERO + MS_IN_DAY * day as u64
+    }
+
+    fn final_timestamp_of_day(day: u16) -> TimestampMillis {
+        Self::day_to_timestamp(day + 1) - 1
     }
 
     fn insurance_cost_for_day(day_index: u8) -> u128 {