+++ frontend/app/rollup.config.mjs
@@ -11,6 +11,7 @@ import inject from "rollup-plugin-inject";
 import json from "@rollup/plugin-json";
 import analyze from "rollup-plugin-analyzer";
 import filesize from "rollup-plugin-filesize";
+import execute from "rollup-plugin-shell";
 import styles from "rollup-styles";
 import alias from "@rollup/plugin-alias";
 import autoprefixer from "autoprefixer";
@@ -201,9 +202,6 @@ export default {
             "import.meta.env.OC_SERVICE_WORKER_PATH": JSON.stringify(
                 process.env.OC_SERVICE_WORKER_PATH,
             ),
-            "import.meta.env.OC_SUSPICIOUS_USERIDS": JSON.stringify(
-                process.env.OC_SUSPICIOUS_USERIDS,
-            ),
             "import.meta.env.OC_BITCOIN_MAINNET_ENABLED": JSON.stringify(
                 process.env.OC_BITCOIN_MAINNET_ENABLED,
             ),
@@ -306,6 +304,10 @@ export default {
             hook: "generateBundle",
         }),
         sourcemapNewline(),
+        execute({
+            commands: [`../../scripts/get-public-key.sh ${process.env.OC_DFX_NETWORK} > ./public/public-key`],
+            hook: "buildStart",
+        })
     ],
     watch: {
         clearScreen: false,