+++ frontend/vite-env.d.ts
@@ -29,7 +29,6 @@ interface ImportMetaEnv {
     readonly OC_SIGN_IN_WITH_ETHEREUM_CANISTER: string;
     readonly OC_SIGN_IN_WITH_SOLANA_CANISTER: string;
     readonly OC_STORAGE_INDEX_CANISTER: string;
-    readonly OC_SUSPICIOUS_USERIDS: string[];
     readonly OC_TENOR_APIKEY: string;
     readonly OC_TRANSLATIONS_CANISTER: string;
     readonly OC_USER_INDEX_CANISTER: string;